+++ CHANGELOG.md
@@ -10,6 +10,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Export genotype with gene variantS search (#6088)
 - igv.js version to 3.7.3 (#6096)
+- Refactor, speedup and add a progress bar to `scout delete variants` cmd (#6094)
+- `scout delete variants` command now accepts an optional `--out-file` where to print a detailed report of the deletion process (#6094)
 ### Fixed
 - Comments' text wrapping in ACMG classifications exported as PDF (#6086)